+++ BuildFeed/Models/ApiModel/NewBuild.cs
@@ -15,10 +15,10 @@ namespace BuildFeed.Models.ApiModel
 
     public class NewBuildObject
     {
-        public byte MajorVersion { get; set; }
-        public byte MinorVersion { get; set; }
-        public ushort Number { get; set; }
-        public ushort? Revision { get; set; }
+        public uint MajorVersion { get; set; }
+        public uint MinorVersion { get; set; }
+        public uint Number { get; set; }
+        public uint? Revision { get; set; }
         public string Lab { get; set; }
         public DateTime? BuildTime { get; set; }
         public LevelOfFlight FlightLevel { get; set; }
